iot-gateway-rpi/requirements.txt
# Raspberry Pi IoT gateway — MQTT bridge + SQLite offline queue
aiomqtt>=2.3.0
python-dotenv>=1.0.0
